Fix #298

+++ mastodon/src/main/java/org/joinmastodon/android/events/StatusCountersUpdatedEvent.java
@@ -4,7 +4,7 @@ import org.joinmastodon.android.model.Status;
 
 public class StatusCountersUpdatedEvent{
 	public String id;
-	public int favorites, reblogs, replies;
+	public long favorites, reblogs, replies;
 	public boolean favorited, reblogged;
 
 	public StatusCountersUpdatedEvent(Status s){
